added joi validation

+++ app.js
@@ -4,8 +4,15 @@ const mongoose = require('mongoose');
 const ejsMate = require('ejs-mate');
 const methodOverride = require('method-override');
 
+// Error handling
+const ExpressError = require('./utils/ExpressError');
+const catchAsync = require('./utils/catchAsync');
+
 // Models
 const Campground = require('./models/campground');
+
+// Validation Schemas
+const { campgroundSchema } = require('./schemas.js');
 
 mongoose.connect('mongodb://localhost:27017/yelp-camp', {
 	useNewUrlParser: true,
@@ -29,6 +36,16 @@ app.set('views', path.join(__dirname, 'views'));
 app.use(express.urlencoded({ extended: true }));
 app.use(methodOverride('_method'));
 
+const validateCampground = (req, res, next) => {
+	const { error } = campgroundSchema.validate(req.body);
+	if (error) {
+		const message = error.details.map((el) => el.message).join(',');
+		throw new ExpressError(message, 400);
+	} else {
+		next();
+	}
+};
+
 app.get('/', (req, res) => {
 	res.render('home');
 });
@@ -42,34 +59,69 @@ app.get('/campgrounds/new', (req, res) => {
 	res.render('campgrounds/new');
 });
 
-app.post('/campgrounds', async (req, res) => {
-	const campground = new Campground(req.body.campground);
-	await campground.save();
-	res.redirect(`/campground/${campground._id}`);
+app.post(
+	'/campgrounds',
+	validateCampground,
+	catchAsync(async (req, res, next) => {
+		// if (!req.body.campground)
+		// 	throw new ExpressError('Invalid Campground Data', 400);
+		const campground = new Campground(req.body.campground);
+		await campground.save();
+		res.redirect(`/campground/${campground._id}`);
+	})
+);
+
+app.get(
+	'/campground/:id',
+	catchAsync(async (req, res) => {
+		const campground = await Campground.findById(req.params.id);
+		res.render('campgrounds/show', { campground });
+	})
+);
+
+app.get(
+	'/campgrounds/:id/edit',
+	catchAsync(async (req, res) => {
+		const campground = await Campground.findById(req.params.id);
+		res.render('campgrounds/edit', { campground });
+	})
+);
+
+app.put(
+	'/campgrounds/:id',
+	validateCampground,
+	catchAsync(async (req, res, next) => {
+		const { id } = req.params;
+
+		if (!id) throw new ExpressError('Invalid Campground ID', 400);
+
+		const campground = await Campground.findByIdAndUpdate(id, {
+			...req.body.campground,
+		});
+		res.redirect(`/campground/${campground._id}`);
+	})
+);
+
+app.delete(
+	'/campgrounds/:id',
+	catchAsync(async (req, res, next) => {
+		const { id } = req.params;
+
+		if (!id) throw new ExpressError('Invalid Campground ID', 400);
+
+		await Campground.findByIdAndDelete(id);
+		res.redirect('/campgrounds');
+	})
+);
+
+app.all('*', (req, res, next) => {
+	next(new ExpressError('Page Not Found', 404));
 });
 
-app.get('/campground/:id', async (req, res) => {
-	const campground = await Campground.findById(req.params.id);
-	res.render('campgrounds/show', { campground });
-});
-
-app.get('/campgrounds/:id/edit', async (req, res) => {
-	const campground = await Campground.findById(req.params.id);
-	res.render('campgrounds/edit', { campground });
-});
-
-app.put('/campgrounds/:id', async (req, res) => {
-	const { id } = req.params;
-	const campground = await Campground.findByIdAndUpdate(id, {
-		...req.body.campground,
-	});
-	res.redirect(`/campground/${campground._id}`);
-});
-
-app.delete('/campgrounds/:id', async (req, res) => {
-	const { id } = req.params;
-	await Campground.findByIdAndDelete(id);
-	res.redirect('/campgrounds');
+app.use((err, req, res, next) => {
+	const { statusCode = 500 } = err;
+	if (!err.message) err.message = 'Oh No! Something Went Wrong';
+	res.status(statusCode).render('error/error', { err });
 });
 
 app.listen(3000, () => {
